@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace YoastSeoForTypo3\YoastSeo\Service;
 
-use TYPO3\CMS\Core\Cache\CacheManager;
+use TYPO3\CMS\Core\Cache\Frontend\FrontendInterface;
 use TYPO3\CMS\Core\Database\ConnectionPool;
 use TYPO3\CMS\Core\Registry;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
@@ -17,28 +17,20 @@ class CrawlerService
     protected const REGISTRY_KEY = 'crawler-%d-%d';
     protected const INDEX_CHUNK = 50;
 
-    /**
-     * @var \TYPO3\CMS\Core\Cache\Frontend\FrontendInterface
-     */
-    protected $cache;
-
-    /**
-     * @var \TYPO3\CMS\Core\Registry
-     */
-    protected $registry;
-
-    public function __construct()
-    {
-        $this->cache = GeneralUtility::makeInstance(CacheManager::class)->getCache('pages');
-        $this->registry = GeneralUtility::makeInstance(Registry::class);
-    }
+    public function __construct(
+        protected FrontendInterface $cache,
+        protected Registry $registry
+    ) {}
 
     public function getAmountOfPages(int $site, int $languageId): int
     {
         return count($this->getPagesToIndex($site, $languageId));
     }
 
-    public function getIndexInformation(int $site, int $languageId, $offset = 0): array
+    /**
+     * @return array{pages: array<int>, current: int, nextOffset: int, total: int}
+     */
+    public function getIndexInformation(int $site, int $languageId, int $offset = 0): array
     {
         $pagesToIndex = $this->getPagesToIndex($site, $languageId);
         $total = count($pagesToIndex);
@@ -51,10 +43,13 @@ class CrawlerService
             'pages' => array_splice($pagesToIndex, $currentOffset, self::INDEX_CHUNK),
             'current' => $currentOffset,
             'nextOffset' => $currentOffset + self::INDEX_CHUNK,
-            'total' => $total
+            'total' => $total,
         ];
     }
 
+    /**
+     * @return array{offset?: int, total?: int}
+     */
     public function getProgressInformation(int $site, int $languageId): array
     {
         return (array)$this->registry->get(
@@ -71,7 +66,7 @@ class CrawlerService
             sprintf(self::REGISTRY_KEY, $site, $languageId),
             [
                 'offset' => $offset,
-                'total' => $total
+                'total' => $total,
             ]
         );
     }
@@ -81,6 +76,9 @@ class CrawlerService
         $this->registry->remove(self::REGISTRY_NAMESPACE, sprintf(self::REGISTRY_KEY, $site, $languageId));
     }
 
+    /**
+     * @return int[]
+     */
     protected function getPagesToIndex(int $site, int $languageId): array
     {
         $cacheIdentifier = 'YoastSeoCrawler' . $site . '-' . $languageId;
@@ -106,11 +104,11 @@ class CrawlerService
                         $queryBuilder->expr()->in(
                             'uid',
                             $treeChunk
-                        )
+                        ),
                     ];
                 }
 
-                $statement = $queryBuilder->select($select)
+                $pages = $queryBuilder->select($select)
                     ->from('pages')
                     ->where(
                         $queryBuilder->expr()->in(
@@ -118,8 +116,7 @@ class CrawlerService
                             YoastUtility::getAllowedDoktypes()
                         ),
                         ...$constraints
-                    )->execute();
-                $pages = GeneralUtility::makeInstance(DbalService::class)->getAllResults($statement);
+                    )->executeQuery()->fetchAllAssociative();
                 $pagesToIndex = array_merge($pagesToIndex, array_column($pages, $select));
             }
         }
