@@ -10,13 +10,16 @@ use TYPO3\CMS\Core\Database\ConnectionPool;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Extbase\Configuration\ConfigurationManager;
 use TYPO3\CMS\Extbase\Configuration\ConfigurationManagerInterface;
-use YoastSeoForTypo3\YoastSeo\Service\DbalService;
 
 class YoastUtility
 {
     protected const COLUMN_NAME_FOCUSKEYWORD = 'tx_yoastseo_focuskeyword';
 
-    public static function getAllowedDoktypes(?array $configuration = null, bool $returnInString = false)
+    /**
+     * @param array<string, mixed>|null $configuration
+     * @return int[]
+     */
+    public static function getAllowedDoktypes(?array $configuration = null): array
     {
         $allowedDoktypes = array_map(function ($doktype) {
             return (int)$doktype;
@@ -32,16 +35,22 @@ class YoastUtility
             }
         }
 
-        $allowedDoktypes = $allowedDoktypes ?: [1];
-
-        if ($returnInString) {
-            return implode(',', $allowedDoktypes);
-        }
-
-        return $allowedDoktypes;
+        return $allowedDoktypes ?: [1];
     }
 
-    public static function snippetPreviewEnabled(int $pageId, array $pageRecord, $pageTs = null): bool
+    /**
+     * @param array<string, mixed>|null $configuration
+     */
+    public static function getAllowedDoktypesList(?array $configuration = null): string
+    {
+        return implode(',', self::getAllowedDoktypes($configuration));
+    }
+
+    /**
+     * @param array<string, mixed> $pageRecord
+     * @param array<string, mixed> $pageTs
+     */
+    public static function snippetPreviewEnabled(int $pageId, array $pageRecord, ?array $pageTs = null): bool
     {
         if (!$GLOBALS['BE_USER'] instanceof BackendUserAuthentication ||
             !$GLOBALS['BE_USER']->check('non_exclude_fields', 'pages:tx_yoastseo_snippetpreview')) {
@@ -79,20 +88,23 @@ class YoastUtility
         return $focusKeyword;
     }
 
+    /**
+     * @return array<string, array{keyword: string, synonyms: string}>
+     */
     public static function getRelatedKeyphrases(string $parentTable, int $parentId): array
     {
         $config = [];
         $queryBuilder = GeneralUtility::makeInstance(ConnectionPool::class)->getQueryBuilderForTable(
             'tx_yoastseo_related_focuskeyword'
         );
-        $statement = $queryBuilder->select('*')
+        $relatedKeyphrases = $queryBuilder->select('*')
             ->from('tx_yoastseo_related_focuskeyword')
             ->where(
                 $queryBuilder->expr()->eq('tablenames', $queryBuilder->createNamedParameter($parentTable)),
                 $queryBuilder->expr()->eq('uid_foreign', $parentId)
             )
-            ->execute();
-        $relatedKeyphrases = GeneralUtility::makeInstance(DbalService::class)->getAllResults($statement);
+            ->executeQuery()
+            ->fetchAllAssociative();
 
         foreach ($relatedKeyphrases as $relatedKeyphrase) {
             $config['rk' . (int)$relatedKeyphrase['uid']] = [
@@ -110,10 +122,10 @@ class YoastUtility
      *
      * You can set development by using TypoScript "module.tx_yoastseo.settings.developmentMode = 1"
      *
-     * @param array|null $configuration
+     * @param array<string, mixed>|null $configuration
      * @return bool
      */
-    public static function inProductionMode(array $configuration = null): bool
+    public static function inProductionMode(?array $configuration = null): bool
     {
         if ($configuration === null) {
             $configuration = self::getTypoScriptConfiguration();
@@ -122,6 +134,9 @@ class YoastUtility
         return !((int)($_ENV['YOAST_DEVELOPMENT_MODE'] ?? 0) === 1 || (int)($configuration['developmentMode'] ?? 0) === 1);
     }
 
+    /**
+     * @return array<string, mixed>
+     */
     protected static function getTypoScriptConfiguration(): array
     {
         $configurationManager = GeneralUtility::makeInstance(ConfigurationManager::class);
@@ -139,7 +154,7 @@ class YoastUtility
      */
     public static function fixAbsoluteUrl(string $url): string
     {
-        if (strpos($url, '/') === 0) {
+        if (str_starts_with($url, '/')) {
             $url = GeneralUtility::getIndpEnv('TYPO3_REQUEST_HOST') . $url;
         }
         return $url;
