@@ -5,57 +5,54 @@ declare(strict_types=1);
 namespace YoastSeoForTypo3\YoastSeo\Controller;
 
 use Psr\Http\Message\ResponseInterface;
-use TYPO3\CMS\Backend\Utility\BackendUtility;
-use TYPO3\CMS\Core\Authentication\BackendUserAuthentication;
-use TYPO3\CMS\Core\Exception\SiteNotFoundException;
-use TYPO3\CMS\Core\Localization\LanguageService;
-use TYPO3\CMS\Core\Site\Entity\Site;
-use TYPO3\CMS\Core\Site\SiteFinder;
-use TYPO3\CMS\Core\Type\Bitmask\Permission;
+use TYPO3\CMS\Core\Pagination\ArrayPaginator;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
-use YoastSeoForTypo3\YoastSeo\Pagination\ArrayPaginator;
+use YoastSeoForTypo3\YoastSeo\Backend\Overview\DataProviderRequest;
+use YoastSeoForTypo3\YoastSeo\Backend\Overview\LanguageMenu\LanguageMenuFactory;
+use YoastSeoForTypo3\YoastSeo\DataProviders\OverviewDataProviderInterface;
 use YoastSeoForTypo3\YoastSeo\Pagination\Pagination;
 
 class OverviewController extends AbstractBackendController
 {
+    /** @var array<string, OverviewDataProviderInterface> */
+    protected array $filters;
+
+    /**
+     * @param iterable<string, OverviewDataProviderInterface> $filters
+     */
+    public function __construct(
+        protected LanguageMenuFactory $languageMenuFactory,
+        iterable $filters
+    ) {
+        foreach ($filters as $key => $dataProvider) {
+            $this->addFilter($key, $dataProvider);
+        }
+    }
+
     public function listAction(int $currentPage = 1): ResponseInterface
     {
         $overviewData = $this->getOverviewData($currentPage) + ['action' => 'list'];
         $moduleTemplate = $this->getModuleTemplate();
         if (!isset($overviewData['pageInformation'])) {
-            return $this->returnResponse($overviewData, $moduleTemplate);
+            return $this->returnResponse('Overview/List', $overviewData, $moduleTemplate);
         }
 
         $moduleTemplate->getDocHeaderComponent()->setMetaInformation($overviewData['pageInformation']);
-        $languageMenuItems = $this->getLanguageMenuItems($overviewData['pageInformation']);
-        if (is_array($languageMenuItems)) {
-            $languageMenu = $moduleTemplate->getDocHeaderComponent()->getMenuRegistry()->makeMenu();
-            $languageMenu->setIdentifier('languageMenu');
-            $languageMenu->setLabel(
-                $this->getLanguageService()->sL('LLL:EXT:core/Resources/Private/Language/locallang_general.xlf:LGL.language')
-            );
-            foreach ($languageMenuItems as $languageMenuItem) {
-                $menuItem = $languageMenu
-                    ->makeMenuItem()
-                    ->setTitle($languageMenuItem['title'])
-                    ->setHref($languageMenuItem['href'])
-                    ->setActive($languageMenuItem['active']);
-                $languageMenu->addMenuItem($menuItem);
-            }
+        $languageMenu = $this->languageMenuFactory->create(
+            $this->request,
+            $moduleTemplate,
+            $overviewData['pageInformation']['uid'] ?? 0
+        );
+        if ($languageMenu !== null) {
             $moduleTemplate->getDocHeaderComponent()->getMenuRegistry()->addMenu($languageMenu);
         }
-        return $this->returnResponse($overviewData, $moduleTemplate);
+
+        return $this->returnResponse('Overview/List', $overviewData, $moduleTemplate);
     }
 
-    public function legacyAction(int $currentPage = 1): void
-    {
-        $overviewData = $this->getOverviewData($currentPage) + ['action' => 'legacy'];
-        if (isset($overviewData['pageInformation'])) {
-            $overviewData['languageMenuItems'] = $this->getLanguageMenuItems($overviewData['pageInformation']);
-        }
-        $this->view->assignMultiple($overviewData);
-    }
-
+    /**
+     * @return array<string, mixed>
+     */
     protected function getOverviewData(int $currentPage): array
     {
         $pageInformation = $this->getPageInformation();
@@ -64,11 +61,12 @@ class OverviewController extends AbstractBackendController
         }
 
         $filters = $this->getAvailableFilters();
-        $activeFilter = $this->getActiveFilter($filters);
-        $currentFilter = $filters[$activeFilter];
+        if ($filters === null) {
+            return [];
+        }
 
-        $params = $this->getParams();
-        $items = GeneralUtility::callUserFunction($currentFilter['dataProvider'], $params, $this) ?: [];
+        $activeFilter = $this->getActiveFilter();
+        $items = $activeFilter->process();
 
         $arrayPaginator = GeneralUtility::makeInstance(
             ArrayPaginator::class,
@@ -85,131 +83,54 @@ class OverviewController extends AbstractBackendController
             'pagination' => $pagination,
             'filters' => $filters,
             'activeFilter' => $activeFilter,
-            'params' => $params,
-            'subtitle' => $this->getLanguageService()->sL($currentFilter['label']),
-            'description' => $currentFilter['description'],
-            'link' => $currentFilter['link'],
+            'params' => $this->getDataProviderRequest(),
         ];
     }
 
+    /**
+     * @return array<string, mixed>|null
+     */
     public function getAvailableFilters(): ?array
     {
-        if (array_key_exists('overview_filters', $GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['yoast_seo'])
-            && is_array($GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['yoast_seo']['overview_filters'])
-        ) {
-            $params = $this->getParams();
-            foreach ($GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['yoast_seo']['overview_filters'] as &$filter) {
-                $filter['numberOfItems'] = (int)GeneralUtility::callUserFunction(
-                    $filter['countProvider'],
-                    $params,
-                    $this
-                );
-            }
-            return (array)$GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['yoast_seo']['overview_filters'];
+        if ($this->filters === []) {
+            return null;
         }
 
-        return null;
+        $dataProviderRequest = $this->getDataProviderRequest();
+        foreach ($this->filters as $dataProvider) {
+            $dataProvider->initialize($dataProviderRequest);
+        }
+
+        return $this->filters;
     }
 
-    protected function getActiveFilter(array $filters): string
+    protected function getActiveFilter(): OverviewDataProviderInterface
     {
-        $activeFilter = '';
+        if ($this->filters === []) {
+            throw new \RuntimeException('No filters available');
+        }
 
         if ($this->request->hasArgument('filter')) {
-            foreach ($filters as $k => $f) {
-                if ($f['key'] === $this->request->getArgument('filter')) {
-                    $activeFilter = $k;
-                    break;
-                }
+            $activeFilter = $this->request->getArgument('filter');
+            if (is_string($activeFilter) && isset($this->filters[$activeFilter])) {
+                return $this->filters[$activeFilter];
             }
         }
 
-        if (empty($activeFilter)) {
-            $activeFilter = key($filters);
-        }
-
-        return (string)$activeFilter;
+        return current($this->filters);
     }
 
-    protected function getLanguageMenuItems(array $pageInformation): ?array
+    protected function getDataProviderRequest(): DataProviderRequest
     {
-        $site = $this->getSite($pageInformation['uid']);
-        if ($site === null) {
-            return null;
-        }
-
-        $languages = $site->getAvailableLanguages($this->getBackendUser());
-        if (count($languages) === 0) {
-            return null;
-        }
-
-        $arguments = $this->getArguments();
-
-        $filter = $arguments['filter'] ?? '';
-        $returnUrl = $arguments['returnUrl'] ?? '';
-        $items = [];
-        foreach ($languages as $language) {
-            $url = $this->uriBuilder
-                ->reset()
-                ->setTargetPageUid($pageInformation['uid'])
-                ->setArguments([
-                    'tx_yoastseo_yoast_yoastseooverview' => [
-                        'filter' => $filter,
-                        'language' => $language->getLanguageId(),
-                        'returnUrl' => $returnUrl,
-                        'controller' => 'Overview'
-                    ]
-                ])
-                ->build();
-            $items[] = [
-                'title' => $language->getTitle(),
-                'href' => $url,
-                'active' => (int)($arguments['language'] ?? 0) === $language->getLanguageId()
-            ];
-        }
-        return $items;
-    }
-
-    protected function getPageInformation(): ?array
-    {
-        $id = GeneralUtility::_GET('id');
-        $pageInformation = BackendUtility::readPageAccess(
-            $id,
-            $this->getBackendUser()->getPagePermsClause(Permission::PAGE_SHOW)
+        return new DataProviderRequest(
+            (int)($this->request->getQueryParams()['id'] ?? 0),
+            (int)($this->request->getQueryParams()['tx_yoastseo_yoast_yoastseooverview']['language'] ?? $this->request->getQueryParams()['language'] ?? 0),
+            'pages'
         );
-        return is_array($pageInformation) ? $pageInformation : null;
     }
 
-    protected function getSite(int $pageUid): ?Site
+    protected function addFilter(string $key, OverviewDataProviderInterface $dataProvider): void
     {
-        $siteFinder = GeneralUtility::makeInstance(SiteFinder::class);
-        try {
-            return $siteFinder->getSiteByPageId($pageUid);
-        } catch (SiteNotFoundException $e) {
-            return null;
-        }
-    }
-
-    protected function getParams(): array
-    {
-        return [
-            'language' => $this->getArguments()['language'] ?? 0,
-            'table' => 'pages'
-        ];
-    }
-
-    protected function getArguments(): array
-    {
-        return $this->request->getArguments()['tx_yoastseo_yoast_yoastseooverview'] ?? $this->request->getArguments();
-    }
-
-    public function getLanguageService(): LanguageService
-    {
-        return $GLOBALS['LANG'];
-    }
-
-    protected function getBackendUser(): BackendUserAuthentication
-    {
-        return $GLOBALS['BE_USER'];
+        $this->filters[$key] = $dataProvider;
     }
 }
