@@ -7,6 +7,7 @@ namespace YoastSeoForTypo3\YoastSeo\DataProviders;
 use TYPO3\CMS\Core\Database\ConnectionPool;
 use TYPO3\CMS\Core\Database\Platform\PlatformInformation;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
+use YoastSeoForTypo3\YoastSeo\Backend\Overview\DataProviderRequest;
 use YoastSeoForTypo3\YoastSeo\Utility\PageAccessUtility;
 
 abstract class AbstractOverviewDataProvider implements OverviewDataProviderInterface
@@ -30,54 +31,62 @@ abstract class AbstractOverviewDataProvider implements OverviewDataProviderInter
         'title',
         'seo_title',
         'tx_yoastseo_score_readability',
-        'tx_yoastseo_score_seo'
+        'tx_yoastseo_score_seo',
     ];
 
-    protected array $callerParams = [];
+    protected DataProviderRequest $dataProviderRequest;
 
-    public function process(array $params): array
+    public function initialize(DataProviderRequest $dataProviderRequest): void
     {
-        $this->callerParams = $params;
-
-        return $this->getRestrictedPagesResults();
+        $this->dataProviderRequest = $dataProviderRequest;
     }
 
-    public function numberOfItems(array $params): int
+    /**
+     * @return array<int, array<string, mixed>>
+     */
+    public function process(): array
     {
-        $this->callerParams = $params;
-
-        return $this->getRestrictedPagesResults(true);
-    }
-
-    protected function getRestrictedPagesResults(bool $returnOnlyCount = false)
-    {
-        $pageIds = PageAccessUtility::getPageIds((int)GeneralUtility::_GET('id'));
-
-        $connection = GeneralUtility::makeInstance(ConnectionPool::class)->getConnectionForTable(self::PAGES_TABLE);
-        $maxBindParameters = PlatformInformation::getMaxBindParameters($connection->getDatabasePlatform());
-
         $pages = [];
-        $pageCount = 0;
-        foreach (array_chunk($pageIds, $maxBindParameters - 10) as $chunk) {
+        foreach (array_chunk($this->getPageIds(), $this->getMaxBindParameters() - 10) as $chunk) {
             $query = $this->getResults($chunk);
             if ($query === null) {
                 continue;
             }
-
-            if ($returnOnlyCount) {
-                $pageCount += $query->rowCount();
-                continue;
-            }
-
-            foreach ($query->fetchAll() as $page) {
+            foreach ($query->fetchAllAssociative() as $page) {
                 $pages[] = $page;
             }
         }
 
-        if ($returnOnlyCount === false) {
-            return $pages;
-        }
+        return $pages;
+    }
 
-        return $pageCount;
+    public function getNumberOfItems(): int
+    {
+        $pageCount = 0;
+        foreach (array_chunk($this->getPageIds(), $this->getMaxBindParameters() - 10) as $chunk) {
+            $query = $this->getResults($chunk);
+            if ($query === null) {
+                continue;
+            }
+            $pageCount += $query->rowCount();
+        }
+        return (int)$pageCount;
+    }
+
+    /**
+     * @return int[]
+     */
+    protected function getPageIds(): array
+    {
+        return PageAccessUtility::getPageIds($this->dataProviderRequest->getId());
+    }
+
+    /**
+     * @return int<999, max>
+     */
+    protected function getMaxBindParameters(): int
+    {
+        $connection = GeneralUtility::makeInstance(ConnectionPool::class)->getConnectionForTable(self::PAGES_TABLE);
+        return max(999, PlatformInformation::getMaxBindParameters($connection->getDatabasePlatform()));
     }
 }
