@@ -8,31 +8,54 @@ use Doctrine\DBAL\Result;
 use TYPO3\CMS\Core\Database\Connection;
 use TYPO3\CMS\Core\Database\ConnectionPool;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
-use YoastSeoForTypo3\YoastSeo\Service\DbalService;
 use YoastSeoForTypo3\YoastSeo\Utility\YoastUtility;
 
 class OrphanedContentDataProvider extends AbstractOverviewDataProvider
 {
+    public function getKey(): string
+    {
+        return 'orphaned';
+    }
+
+    public function getLabel(): string
+    {
+        return 'LLL:EXT:yoast_seo/Resources/Private/Language/BackendModuleOverview.xlf:orphanedContent';
+    }
+
+    public function getDescription(): string
+    {
+        return 'LLL:EXT:yoast_seo/Resources/Private/Language/BackendModuleOverview.xlf:orphanedContent.description';
+    }
+
+    public function getLink(): ?string
+    {
+        return 'https://yoa.st/1ja';
+    }
+
+    /** @var int[] */
     protected array $referencedPages = [];
 
+    /**
+     * @param int[] $pageIds
+     */
     public function getResults(array $pageIds = []): ?Result
     {
         if ($this->referencedPages === []) {
-            $this->setReferencedPages();
+            $this->referencedPages = $this->getReferencedPages();
         }
 
         $qb = GeneralUtility::makeInstance(ConnectionPool::class)->getQueryBuilderForTable('pages');
 
         $constraints = [
             $qb->expr()->in('doktype', YoastUtility::getAllowedDoktypes()),
-            $qb->expr()->eq('sys_language_uid', (int)$this->callerParams['language'])
+            $qb->expr()->eq('sys_language_uid', $this->dataProviderRequest->getLanguage())
         ];
         if (count($this->referencedPages) > 0) {
             $constraints[] = $qb->expr()->notIn('uid', $this->referencedPages);
         }
         if (count($pageIds) > 0) {
             $constraints[] = $qb->expr()->in(
-                (int)$this->callerParams['language'] > 0 ? $GLOBALS['TCA']['pages']['ctrl']['transOrigPointerField'] : 'uid',
+                $this->dataProviderRequest->getLanguage() > 0 ? $GLOBALS['TCA']['pages']['ctrl']['transOrigPointerField'] : 'uid',
                 $pageIds
             );
         }
@@ -40,10 +63,13 @@ class OrphanedContentDataProvider extends AbstractOverviewDataProvider
         return $qb->select('*')
             ->from('pages')
             ->where(...$constraints)
-            ->execute();
+            ->executeQuery();
     }
 
-    protected function setReferencedPages(): void
+    /**
+     * @return int[]
+     */
+    protected function getReferencedPages(): array
     {
         $qb = GeneralUtility::makeInstance(ConnectionPool::class)->getQueryBuilderForTable('sys_refindex');
 
@@ -58,16 +84,13 @@ class OrphanedContentDataProvider extends AbstractOverviewDataProvider
             )
         ];
 
-        $statement = $qb->select('ref_uid')
+        $references = $qb->select('ref_uid')
             ->from('sys_refindex')
             ->where(...$constraints)
             ->groupBy('ref_uid')
-            ->execute();
+            ->executeQuery()
+            ->fetchAllAssociative();
 
-        $refs = GeneralUtility::makeInstance(DbalService::class)->getAllResults($statement);
-
-        foreach ($refs as $ref) {
-            $this->referencedPages[] = $ref['ref_uid'];
-        }
+        return array_column($references, 'ref_uid');
     }
 }
