@@ -5,12 +5,14 @@ declare(strict_types=1);
 namespace YoastSeoForTypo3\YoastSeo\Form\Element;
 
 use TYPO3\CMS\Backend\Form\AbstractNode;
-use TYPO3\CMS\Backend\Form\NodeFactory;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Fluid\View\StandaloneView;
 
 class Cornerstone extends AbstractNode
 {
+    /**
+     * @return array<string, mixed>
+     */
     public function render(): array
     {
         $resultArray = $this->initializeResultArray();
