@@ -29,6 +29,9 @@ class SnippetPreview extends AbstractNode
 
     protected UrlService $urlService;
 
+    /**
+     * @return array<string, mixed>
+     */
     public function render(): array
     {
         $this->initialize();
@@ -100,27 +103,15 @@ class SnippetPreview extends AbstractNode
     {
         $this->urlService = GeneralUtility::makeInstance(UrlService::class);
 
-        if (array_key_exists('titleField', (array)$this->data['parameterArray']['fieldConf']['config']['settings']) &&
-            $this->data['parameterArray']['fieldConf']['config']['settings']['titleField']
-        ) {
+        if (!empty($this->data['parameterArray']['fieldConf']['config']['settings']['titleField'] ?? '')) {
             $this->titleField = $this->data['parameterArray']['fieldConf']['config']['settings']['titleField'];
         }
 
-        if (array_key_exists(
-                'pageTitleField',
-                (array)$this->data['parameterArray']['fieldConf']['config']['settings']
-            ) &&
-            $this->data['parameterArray']['fieldConf']['config']['settings']['pageTitleField']
-        ) {
+        if (!empty($this->data['parameterArray']['fieldConf']['config']['settings']['pageTitleField'] ?? '')) {
             $this->pageTitleField = $this->data['parameterArray']['fieldConf']['config']['settings']['pageTitleField'];
         }
 
-        if (array_key_exists(
-                'descriptionField',
-                (array)$this->data['parameterArray']['fieldConf']['config']['settings']
-            )
-            && $this->data['parameterArray']['fieldConf']['config']['settings']['descriptionField']
-        ) {
+        if (!empty($this->data['parameterArray']['fieldConf']['config']['settings']['descriptionField'] ?? '')) {
             $this->descriptionField =
                 $this->data['parameterArray']['fieldConf']['config']['settings']['descriptionField'];
         }
@@ -199,7 +190,7 @@ class SnippetPreview extends AbstractNode
         // map record data to GET parameters
         if (isset($previewConfiguration['fieldToParameterMap.'])) {
             foreach ($previewConfiguration['fieldToParameterMap.'] as $field => $parameterName) {
-                $value = $recordArray[$field];
+                $value = $recordArray[$field] ?? '';
                 if ($field === 'uid') {
                     $value = $recordId;
                 }
@@ -232,6 +223,9 @@ class SnippetPreview extends AbstractNode
         return $this->urlService->getPreviewUrl($previewPageId, $languageId, $additionalParamsForUrl);
     }
 
+    /**
+     * @param array<string, mixed> $previewConfiguration
+     */
     protected function getPreviewPageId(int $currentPageId, array $previewConfiguration): int
     {
         // find the right preview page id
@@ -271,8 +265,8 @@ class SnippetPreview extends AbstractNode
      * The result can be used to create a query string with
      * GeneralUtility::implodeArrayForUrl().
      *
-     * @param array $parameters Should be an empty array by default
-     * @param array $typoScript The TypoScript configuration
+     * @param array<string, mixed> $parameters Should be an empty array by default
+     * @param array<string, mixed> $typoScript The TypoScript configuration
      */
     protected function parseAdditionalGetParameters(
         array &$parameters,
