@@ -5,20 +5,15 @@ declare(strict_types=1);
 namespace YoastSeoForTypo3\YoastSeo\MetaTag;
 
 use TYPO3\CMS\Core\Utility\GeneralUtility;
+use YoastSeoForTypo3\YoastSeo\MetaTag\Generator\GeneratorInterface;
 use YoastSeoForTypo3\YoastSeo\Record\Record;
 use YoastSeoForTypo3\YoastSeo\Record\RecordService;
 
 class RecordMetaTagGenerator
 {
-    protected RecordService $recordService;
-
-    public function __construct(RecordService $recordService = null)
-    {
-        if ($recordService === null) {
-            $recordService = GeneralUtility::makeInstance(RecordService::class);
-        }
-        $this->recordService = $recordService;
-    }
+    public function __construct(
+        protected RecordService $recordService
+    ) {}
 
     public function generate(): void
     {
@@ -27,8 +22,12 @@ class RecordMetaTagGenerator
             return;
         }
 
+        /** @var class-string $generatorClass */
         foreach ($GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['yoast_seo']['recordMetaTags'] ?? [] as $generatorClass) {
-            GeneralUtility::makeInstance($generatorClass)->generate($activeRecord);
+            $generator = GeneralUtility::makeInstance($generatorClass);
+            if ($generator instanceof GeneratorInterface) {
+                $generator->generate($activeRecord);
+            }
         }
     }
 }
