@@ -20,13 +20,13 @@ class ProminentWordsService
     protected string $table;
     protected int $languageId;
 
-    protected ConnectionPool $connectionPool;
+    public function __construct(
+        protected ConnectionPool $connectionPool
+    ) {}
 
-    public function __construct(ConnectionPool $connectionPool)
-    {
-        $this->connectionPool = $connectionPool;
-    }
-
+    /**
+     * @param array<string, int|string> $prominentWords
+     */
     public function saveProminentWords(
         int $uid,
         ?int $pid,
@@ -56,6 +56,9 @@ class ProminentWordsService
         $this->createNewWords($prominentWords);
     }
 
+    /**
+     * @param array{uid: int, weight: int} $oldWord
+     */
     protected function updateIfWeightHasChanged(array $oldWord, int $weight): void
     {
         if ((int)$oldWord['weight'] === $weight) {
@@ -68,9 +71,12 @@ class ProminentWordsService
                 $queryBuilder->expr()->eq('uid', $oldWord['uid'])
             )
             ->setMaxResults(1)
-            ->execute();
+            ->executeStatement();
     }
 
+    /**
+     * @param int[] $wordsToDelete
+     */
     protected function deleteProminentWords(array $wordsToDelete): void
     {
         if ($wordsToDelete === []) {
@@ -85,13 +91,16 @@ class ProminentWordsService
                     $queryBuilder->createNamedParameter($wordsToDelete, Connection::PARAM_INT_ARRAY)
                 )
             )
-            ->execute();
+            ->executeStatement();
     }
 
+    /**
+     * @param array<string, int|string> $prominentWords
+     */
     protected function createNewWords(array $prominentWords): void
     {
         $site = $this->getSiteRootPageId();
-        foreach ($prominentWords ?? [] as $word => $weight) {
+        foreach ($prominentWords as $word => $weight) {
             $data = [
                 'pid' => $this->table === 'pages' ? $this->uid : $this->pid,
                 'sys_language_uid' => $this->languageId,
@@ -99,33 +108,35 @@ class ProminentWordsService
                 'site' => $site,
                 'tablenames' => $this->table,
                 'stem' => $word,
-                'weight' => $weight
+                'weight' => (int)$weight
             ];
             $this->connectionPool->getConnectionForTable(self::PROMINENT_WORDS_TABLE)
                 ->insert(self::PROMINENT_WORDS_TABLE, $data);
         }
     }
 
+    /**
+     * @return array<array{uid: int, stem: string, weight: int}>
+     */
     protected function getOldWords(): array
     {
         $queryBuilder = $this->getQueryBuilder();
-        $statement = $queryBuilder->select('uid', 'stem', 'weight')
+        $queryBuilder->select('uid', 'stem', 'weight')
             ->from(self::PROMINENT_WORDS_TABLE)
             ->where(
                 $queryBuilder->expr()->eq('uid_foreign', $this->uid),
                 $queryBuilder->expr()->eq('tablenames', $queryBuilder->createNamedParameter($this->table)),
                 $queryBuilder->expr()->eq('sys_language_uid', $this->languageId)
-            )
-            ->execute();
-        return GeneralUtility::makeInstance(DbalService::class)->getAllResults($statement);
+            );
+        /** @var array<array{uid: int, stem: string, weight: int}> $oldWords */
+        $oldWords = $queryBuilder->executeQuery()->fetchAllAssociative();
+        return $oldWords;
     }
 
     protected function getPidForRecord(int $uid, string $table): int
     {
         $connection = $this->connectionPool->getConnectionForTable($table);
-        $record = GeneralUtility::makeInstance(DbalService::class)->getSingleResult(
-            $connection->select(['pid'], $table, ['uid' => $uid])
-        );
+        $record = $connection->select(['pid'], $table, ['uid' => $uid])->fetchAssociative();
         return (int)($record['pid'] ?? 0);
     }
 
