@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace YoastSeoForTypo3\YoastSeo\Service;
 
+use TYPO3\CMS\Core\Information\Typo3Version;
 use TYPO3\CMS\Core\Page\PageRenderer;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use YoastSeoForTypo3\YoastSeo\Utility\JavascriptUtility;
@@ -12,20 +13,17 @@ use YoastSeoForTypo3\YoastSeo\Utility\PathUtility;
 
 class SnippetPreviewService
 {
-    protected UrlService $urlService;
-    protected PageRenderer $pageRenderer;
-    protected LocaleService $localeService;
-
     public function __construct(
-        UrlService $urlService,
-        PageRenderer $pageRenderer,
-        LocaleService $localeService
+        protected UrlService $urlService,
+        protected PageRenderer $pageRenderer,
+        protected LocaleService $localeService
     ) {
-        $this->urlService = $urlService;
-        $this->pageRenderer = $pageRenderer;
-        $this->localeService = $localeService;
     }
 
+    /**
+     * @param array<string, mixed> $currentData
+     * @param array<string, mixed> $additionalConfiguration
+     */
     public function buildSnippetPreview(
         string $previewUrl,
         array $currentData,
@@ -53,7 +51,15 @@ class SnippetPreviewService
 
         JavascriptUtility::loadJavascript($this->pageRenderer);
 
-        $this->pageRenderer->loadRequireJsModule('TYPO3/CMS/YoastSeo/yoastModal');
+        if (GeneralUtility::makeInstance(Typo3Version::class)->getMajorVersion() >= 13) {
+            $this->pageRenderer->loadJavaScriptModule(
+                '@yoast/yoast-seo-for-typo3/yoastModalEs6.js',
+            );
+        } else {
+            $this->pageRenderer->loadRequireJsModule(
+                'TYPO3/CMS/YoastSeo/yoastModal',
+            );
+        }
         $this->pageRenderer->addCssFile('EXT:yoast_seo/Resources/Public/CSS/yoast.min.css');
     }
 }
