@@ -5,8 +5,6 @@ declare(strict_types=1);
 namespace YoastSeoForTypo3\YoastSeo\StructuredData;
 
 use Psr\Http\Message\ServerRequestInterface;
-use TYPO3\CMS\Core\Cache\CacheManager;
-use TYPO3\CMS\Core\Cache\Exception\NoSuchCacheException;
 use TYPO3\CMS\Core\Cache\Frontend\FrontendInterface;
 use TYPO3\CMS\Core\Http\ApplicationType;
 use TYPO3\CMS\Core\Service\DependencyOrderingService;
@@ -21,33 +19,32 @@ use TYPO3\CMS\Frontend\Controller\TypoScriptFrontendController;
  */
 class StructuredDataProviderManager implements SingletonInterface
 {
-    protected FrontendInterface $pageCache;
-    protected string $sourceComment = '';
+    public function __construct(
+        protected FrontendInterface $pageCache
+    ) {}
 
-    public function __construct()
+    /**
+     * @param array<string, mixed> $params
+     */
+    public function render(array &$params, object $pObj): void
     {
-        $this->initCaches();
-        $this->sourceComment = '<!-- This site is optimized with the Yoast SEO for TYPO3 plugin - https://yoast.com/typo3-extensions-seo/ -->';
-    }
-
-    public function render(array &$params, $pObj): void
-    {
-        if (!($GLOBALS['TYPO3_REQUEST'] ?? null) instanceof ServerRequestInterface
-            || !ApplicationType::fromRequest($GLOBALS['TYPO3_REQUEST'])->isFrontend()
-        ) {
+        if (!$this->isFrontendRequest()) {
             return;
         }
 
         $data = $this->getStructuredData();
 
-        $params['headerData']['StructuredDataManager'] = $this->sourceComment .
-            PHP_EOL . $this->buildJsonLdBlob($data);
+        $params['headerData']['StructuredDataManager']
+            = $this->getSourceComment() . PHP_EOL . $this->buildJsonLdBlob($data);
     }
 
+    /**
+     * @param array<string, array<string, mixed>> $src
+     */
     protected function buildJsonLdBlob(array $src): string
     {
         $data = [];
-        foreach ($src as $provider => $dataItems) {
+        foreach ($src as $dataItems) {
             foreach ($dataItems as $item) {
                 $data[] = $item;
             }
@@ -60,50 +57,58 @@ class StructuredDataProviderManager implements SingletonInterface
         return '<script type="application/ld+json">' . json_encode($data) . '</script>';
     }
 
+    /**
+     * @return array<string, array<string, mixed>>
+     */
     public function getStructuredData(): array
     {
         $structuredData = [];
-
-        $structuredDataProviders = $this->getStructuredDataProviderConfiguration();
-        $structuredDataProviders = $this->setProviderOrder($structuredDataProviders);
-
-        $orderedStructuredDataProviders = GeneralUtility::makeInstance(DependencyOrderingService::class)
-            ->orderByDependencies($structuredDataProviders);
-
-        foreach ($orderedStructuredDataProviders as $provider => $configuration) {
+        foreach ($this->getOrderedStructuredDataProviders() as $provider => $configuration) {
             $cacheIdentifier = $this->getTypoScriptFrontendController()->newHash . '-structured-data-' . $provider;
-            if ($this->pageCache instanceof FrontendInterface &&
-                $data = $this->pageCache->get($cacheIdentifier)
-            ) {
-                if (!empty($data)) {
+            if ($this->pageCache instanceof FrontendInterface) {
+                $data = $this->pageCache->get($cacheIdentifier);
+                if ($data !== false) {
                     $structuredData[$provider] = $data;
+                    continue;
                 }
+            }
+            $structuredDataProviderObject = $this->getStructuredDataProviderObject($configuration);
+            if ($structuredDataProviderObject === null) {
                 continue;
             }
-            if (class_exists($configuration['provider'])
-                && is_subclass_of($configuration['provider'], StructuredDataProviderInterface::class)) {
-                /** @var StructuredDataProviderInterface $structuredDataProviderObject */
-                $structuredDataProviderObject = GeneralUtility::makeInstance($configuration['provider']);
-                if (method_exists($structuredDataProviderObject, 'setConfiguration')) {
-                    $structuredDataProviderObject->setConfiguration($configuration);
-                }
 
-                if ($data = $structuredDataProviderObject->getData()) {
-                    $this->pageCache->set(
-                        $cacheIdentifier,
-                        $data,
-                        ['pageId_' . $this->getTypoScriptFrontendController()->page['uid']],
-                        $this->getTypoScriptFrontendController()->get_cache_timeout()
-                    );
-                }
+            if ($data = $structuredDataProviderObject->getData()) {
+                $this->pageCache->set(
+                    $cacheIdentifier,
+                    $data,
+                    ['pageId_' . ($this->getTypoScriptFrontendController()->page['uid'] ?? $this->getTypoScriptFrontendController()->id)],
+                    // TODO: Fix this call, protected method since v13
+                    //$this->getTypoScriptFrontendController()->get_cache_timeout()
+                );
+            }
 
-                if (!empty($data)) {
-                    $structuredData[$provider] = $data;
-                }
+            if (!empty($data)) {
+                $structuredData[$provider] = $data;
             }
         }
 
         return $structuredData;
+    }
+
+    /**
+     * @param array<string, mixed> $configuration
+     */
+    protected function getStructuredDataProviderObject(array $configuration): StructuredDataProviderInterface|null
+    {
+        if (!class_exists($configuration['provider']) || !is_subclass_of($configuration['provider'], StructuredDataProviderInterface::class)) {
+            return null;
+        }
+
+        $providerObject = GeneralUtility::makeInstance($configuration['provider']);
+        if (method_exists($providerObject, 'setConfiguration')) {
+            $providerObject->setConfiguration($configuration);
+        }
+        return $providerObject;
     }
 
     private function getTypoScriptFrontendController(): TypoScriptFrontendController
@@ -111,6 +116,21 @@ class StructuredDataProviderManager implements SingletonInterface
         return $GLOBALS['TSFE'];
     }
 
+    /**
+     * @return array<string, array<string, mixed>>
+     */
+    private function getOrderedStructuredDataProviders(): array
+    {
+        $structuredDataProviders = $this->getStructuredDataProviderConfiguration();
+        $structuredDataProviders = $this->setProviderOrder($structuredDataProviders);
+
+        return GeneralUtility::makeInstance(DependencyOrderingService::class)
+            ->orderByDependencies($structuredDataProviders);
+    }
+
+    /**
+     * @return array<string, array<string, mixed>>
+     */
     private function getStructuredDataProviderConfiguration(): array
     {
         $typoscriptService = GeneralUtility::makeInstance(TypoScriptService::class);
@@ -121,41 +141,49 @@ class StructuredDataProviderManager implements SingletonInterface
         return $config['structuredData']['providers'] ?? [];
     }
 
-    protected function initCaches()
-    {
-        try {
-            $this->pageCache = GeneralUtility::makeInstance(CacheManager::class)->getCache(
-                'pages'
-            );
-        } catch (NoSuchCacheException $e) {
-            // @ignoreException
-        }
-    }
-
+    /**
+     * @param array<string, array<string, mixed>> $orderInformation
+     * @return array<string, array<string, mixed>>
+     */
     protected function setProviderOrder(array $orderInformation): array
     {
         foreach ($orderInformation as $provider => &$configuration) {
             if (isset($configuration['before'])) {
-                if (is_string($configuration['before'])) {
-                    $configuration['before'] = GeneralUtility::trimExplode(',', $configuration['before'], true);
-                } elseif (!is_array($configuration['before'])) {
-                    throw new \UnexpectedValueException(
-                        'The specified "before" order configuration for provider "' . $provider . '" is invalid.',
-                        1551014599
-                    );
-                }
+                $configuration['before'] = $this->getOrderConfiguration($provider, $configuration, 'before');
             }
             if (isset($configuration['after'])) {
-                if (is_string($configuration['after'])) {
-                    $configuration['after'] = GeneralUtility::trimExplode(',', $configuration['after'], true);
-                } elseif (!is_array($configuration['after'])) {
-                    throw new \UnexpectedValueException(
-                        'The specified "after" order configuration for provider "' . $provider . '" is invalid.',
-                        1551014604
-                    );
-                }
+                $configuration['after'] = $this->getOrderConfiguration($provider, $configuration, 'after');
             }
         }
         return $orderInformation;
+    }
+
+    /**
+     * @param array<string, mixed> $configuration
+     * @return string[]
+     */
+    private function getOrderConfiguration(string $provider, array $configuration, string $key): array
+    {
+        if (is_string($configuration[$key])) {
+            return GeneralUtility::trimExplode(',', $configuration[$key], true);
+        }
+        if (!is_array($configuration[$key])) {
+            throw new \UnexpectedValueException(
+                'The specified "' . $key . '" order configuration for provider "' . $provider . '" is invalid.',
+                1551014599
+            );
+        }
+        return $configuration[$key];
+    }
+
+    protected function isFrontendRequest(): bool
+    {
+        return ($GLOBALS['TYPO3_REQUEST'] ?? null) instanceof ServerRequestInterface
+            && ApplicationType::fromRequest($GLOBALS['TYPO3_REQUEST'])->isFrontend();
+    }
+
+    protected function getSourceComment(): string
+    {
+        return '<!-- This site is optimized with the Yoast SEO for TYPO3 plugin - https://yoast.com/typo3-extensions-seo/ -->';
     }
 }
