@@ -9,19 +9,17 @@ use TYPO3\CMS\Core\Utility\ArrayUtility;
 
 class JsonConfigUtility implements SingletonInterface
 {
+    /** @var array<string, mixed> */
     protected array $config = [];
 
     /**
-     * @param array $config
+     * @param array<string, mixed> $config
      */
-    public function addConfig(array $config)
+    public function addConfig(array $config): void
     {
         ArrayUtility::mergeRecursiveWithOverrule($this->config, $config, true, false);
     }
 
-    /**
-     * @return string
-     */
     public function render(): string
     {
         return 'const YoastConfig = ' . json_encode($this->config) . ';';
