@@ -5,17 +5,21 @@ declare(strict_types=1);
 namespace YoastSeoForTypo3\YoastSeo\Utility;
 
 use TYPO3\CMS\Core\Authentication\BackendUserAuthentication;
+use TYPO3\CMS\Core\Database\Connection;
 use TYPO3\CMS\Core\Database\ConnectionPool;
 use TYPO3\CMS\Core\Database\Query\QueryHelper;
 use TYPO3\CMS\Core\Database\Query\Restriction\DeletedRestriction;
 use TYPO3\CMS\Core\Type\Bitmask\Permission;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
-use YoastSeoForTypo3\YoastSeo\Service\DbalService;
 
 class PageAccessUtility
 {
+    /** @var int[] */
     protected static array $cache = [];
 
+    /**
+     * @return int[]
+     */
     public static function getPageIds(int $pid): array
     {
         if (self::$cache !== []) {
@@ -33,52 +37,39 @@ class PageAccessUtility
 
     /**
      * Copied from EXT:core, removed in v12
-     *
-     * @param $id
-     * @param $depth
-     * @param $begin
-     * @param $permClause
-     * @return string
      */
     protected static function getTreeList(int $id, int $depth, int $begin = 0, string $permClause = ''): string
     {
-        $depth = (int)$depth;
-        $begin = (int)$begin;
-        $id = (int)$id;
         if ($id < 0) {
             $id = abs($id);
         }
-        if ($begin === 0) {
-            $theList = $id;
-        } else {
-            $theList = '';
+        $theList = $begin === 0 ? $id : '';
+        if (!($id && $depth > 0)) {
+            return (string)$theList;
         }
-        if ($id && $depth > 0) {
-            $queryBuilder = GeneralUtility::makeInstance(ConnectionPool::class)->getQueryBuilderForTable('pages');
-            $queryBuilder->getRestrictions()->removeAll()->add(GeneralUtility::makeInstance(DeletedRestriction::class));
-            $queryBuilder->select('uid')
-                ->from('pages')
-                ->where(
-                    $queryBuilder->expr()->eq('pid', $queryBuilder->createNamedParameter($id, \PDO::PARAM_INT)),
-                    $queryBuilder->expr()->eq('sys_language_uid', 0)
-                )
-                ->orderBy('uid');
-            if ($permClause !== '') {
-                $queryBuilder->andWhere(QueryHelper::stripLogicalOperatorPrefix($permClause));
+
+        $queryBuilder = GeneralUtility::makeInstance(ConnectionPool::class)->getQueryBuilderForTable('pages');
+        $queryBuilder->getRestrictions()->removeAll()->add(GeneralUtility::makeInstance(DeletedRestriction::class));
+        $queryBuilder->select('uid')
+            ->from('pages')
+            ->where(
+                $queryBuilder->expr()->eq('pid', $queryBuilder->createNamedParameter($id, Connection::PARAM_INT)),
+                $queryBuilder->expr()->eq('sys_language_uid', 0)
+            )
+            ->orderBy('uid');
+        if ($permClause !== '') {
+            $queryBuilder->andWhere(QueryHelper::stripLogicalOperatorPrefix($permClause));
+        }
+        foreach ($queryBuilder->executeQuery()->fetchAllAssociative() as $row) {
+            if ($begin <= 0) {
+                $theList .= ',' . $row['uid'];
             }
-            $statement = $queryBuilder->execute();
-            $rows = GeneralUtility::makeInstance(DbalService::class)->getAllResults($statement);
-            foreach ($rows as $row) {
-                if ($begin <= 0) {
-                    $theList .= ',' . $row['uid'];
+            if ($depth > 1) {
+                $theSubList = self::getTreeList($row['uid'], $depth - 1, $begin - 1, $permClause);
+                if (!empty($theList) && !empty($theSubList) && ($theSubList[0] !== ',')) {
+                    $theList .= ',';
                 }
-                if ($depth > 1) {
-                    $theSubList = self::getTreeList($row['uid'], $depth - 1, $begin - 1, $permClause);
-                    if (!empty($theList) && !empty($theSubList) && ($theSubList[0] !== ',')) {
-                        $theList .= ',';
-                    }
-                    $theList .= $theSubList;
-                }
+                $theList .= $theSubList;
             }
         }
         return (string)$theList;
