@@ -10,9 +10,6 @@ use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3Fluid\Fluid\Core\Rendering\RenderingContextInterface;
 use TYPO3Fluid\Fluid\Core\ViewHelper\AbstractViewHelper;
 
-/**
- * Class RecordIconViewHelper
- */
 class RecordIconViewHelper extends AbstractViewHelper
 {
     protected $escapeOutput = false;
@@ -25,10 +22,7 @@ class RecordIconViewHelper extends AbstractViewHelper
     }
 
     /**
-     * @param array $arguments
-     * @param \Closure $renderChildrenClosure
-     * @param \TYPO3Fluid\Fluid\Core\Rendering\RenderingContextInterface $renderingContext
-     * @return string
+     * @param array<string, mixed> $arguments
      */
     public static function renderStatic(
         array $arguments,
